cmake for Win32/Linux; fix Win32 warnings; fix C++11 build; keep up with bazel incompatibilities.
Details:
- BUILD  (and bazel/BUILD.bazel)
        Avoid
                includes = ["."],
        which bazel doesn't like when the WORKSPACE file is in the same
        directory as the BUILD file.
        Also, allow use of platform_c++11_os.h files.
- CMakeLists.txt
        Add a cmake config file.  Works on Linux and Win32 so far.
- builds/x86_64.linux.c++11/dependfile
        Should be empty until filled in by make depend.
- builds/x86_64.macos.c++11/Makefile
- builds/x86_64.macos.c++11/dependfile
- builds/x86_64.win32.c++11/Makefile
- builds/x86_64.win32.c++11/dependfile
        Allow the C++11 build on Windows and Macos.
- internal/sem.h
        Some C++11 data strcutures are huge, so need more space to hold them.
- platform/c++11/atomic.h
        Use platform-independent INLINE instead of __inline__
- platform/c++11/platform.h
        Add a level of indirection when simulating pthread synchronization
        types with C++11 types, because C++ destructors interfere with
        C semantics.
        Bug fix in nsync_pthread_cond_wait() and nsync_pthread_cond_timedwait_();
        use std::adopt_lock rather than std::defer_lock.
        Add platform_c++11_os.h files to allow MacOS and Win32 to
        augment the C++11 libraries with OS-specific ones.
- platform/c++11/src/time_rep_timespec.cc
        Make conversions explicit to avoid warning.
- platform/macos/platform_c++11_os.h
- platform/posix/platform_c++11_os.h
        Provide OS-specific things needed when using the C++11 build.
        The posix file defineds nothing, but the Win32 and MacOS
        files need some things that not available by default on those systems.
- platform/win32/platform.h
        Avoid warning about tmpfile().
- platform/win32/platform_c++11_os.h
        Provide OS-specific things needed when using the C++11 build.
        The posix file defineds nothing, but the Win32 and MacOS
        files need some things that not available by default on those systems.
- platform/win32/src/nanosleep.c
  platform/win32/src/nsync_semaphore_win32.c
  platform/win32/src/pthread_cond_timedwait_win32.c
- testing/cv_test.c
  testing/dll_test.c
        Avoid strcpy() to avoid warning.
- testing/mu_starvation_test.c
        Make conversions explicit to avoid warning.
- testing/pingpong.c
        Bug fix: destruct done_cond.
- testing/testing.c
  testing/time_extra.c
        Include compiler.h to get INLINE, now needed by platform/c++11/atomic.h
        Make conversions explicit to avoid warning.
        Avoid strncpy() to avoid warning.

+++ platform/c++11/src/time_rep_timespec.cc
@@ -59,7 +59,7 @@ nsync_time nsync_from_time_point_ (nsync_cpp_time_point_ tp) {
 	struct timespec ts;
 	memset (&ts, 0, sizeof (ts));
 	ts.tv_sec = ns / NSYNC_NS_IN_S_;
-	ts.tv_nsec = ns - ts.tv_sec * NSYNC_NS_IN_S_;
+	ts.tv_nsec = (long) (ns - ts.tv_sec * NSYNC_NS_IN_S_);
 	return (ts);
 }
